requirements.txt
urllib3<1.26.0
pyyaml>=6.0
pyarrow>=6.0.0
pandas>=1.3.4
signalrcore==0.9.2
matlink-gpapi>=0.4.4.5
requests>=2.26.0 

findspark>=1.4.2
pyspark>=3.2.0 